backend/requirements.txt
fastapi>=0.111,<1.0
uvicorn[standard]>=0.30,<1.0
pydantic>=2.7,<3.0
pydantic-settings>=2.3,<3.0
sqlalchemy>=2.0,<3.0
requests>=2.32,<3.0
beautifulsoup4>=4.12,<5.0
lxml>=5.2,<6.0
sentence-transformers>=3.0,<4.0
numpy>=1.26,<3.0
pdfplumber>=0.11,<1.0
python-docx>=1.1,<2.0
python-multipart>=0.0.9,<1.0
playwright>=1.45,<2.0
pytest>=8.0,<9.0
httpx>=0.27,<1.0

